+++ gluecodium/src/test/resources/smoke/visibility_attribute/output/android/com/example/smoke/PublicInterface.java
@@ -1,0 +1,14 @@
+/*
+ *
+ */
+package com.example.smoke;
+import android.support.annotation.NonNull;
+public interface PublicInterface {
+    static final class InternalStruct {
+        @NonNull
+        public PublicClass.InternalStruct fieldOfInternalType;
+        InternalStruct(@NonNull final PublicClass.InternalStruct fieldOfInternalType) {
+            this.fieldOfInternalType = fieldOfInternalType;
+        }
+    }
+}