+++ gluecodium/src/test/resources/smoke/visibility_attribute/output/android/com/example/smoke/PublicTypeCollection.java
@@ -1,0 +1,15 @@
+/*
+ *
+ */
+package com.example.smoke;
+import android.support.annotation.NonNull;
+public final class PublicTypeCollection {
+    static final class InternalStruct {
+        @NonNull
+        String stringField;
+        InternalStruct(@NonNull final String stringField) {
+            this.stringField = stringField;
+        }
+        public native void fooBar();
+    }
+}