+++ lime-runtime/src/main/java/com/here/gluecodium/model/lime/LimeAttributeType.kt
@@ -35,6 +35,7 @@ enum class LimeAttributeType(
     ENABLE_IF("EnableIf", LimeAttributeValueType.TAG),
     EQUATABLE("Equatable"),
     IMMUTABLE("Immutable"),
+    INTERNAL("Internal"),
     NO_CACHE("NoCache"),
     OPTIMIZED("Optimized"),
     SERIALIZABLE("Serializable"),